refactor: move components of input group into their own files
