+++ src/components/InputGroup/InputGroup.tsx
@@ -1,13 +1,6 @@
 import * as React from 'react';
 
-import { cva } from 'class-variance-authority';
-import type { VariantProps } from 'class-variance-authority';
-
 import { cn } from '#utils';
-
-import { Button } from '../Button/Button.tsx';
-import { Input } from '../Input/Input.tsx';
-import { TextArea } from '../TextArea/TextArea.tsx';
 
 const InputGroup = ({ className, ...props }: React.ComponentProps<'div'>) => {
   return (
@@ -23,108 +16,4 @@ const InputGroup = ({ className, ...props }: React.ComponentProps<'div'>) => {
   );
 };
 
-const inputGroupAddonVariants = cva(
-  "text-muted-foreground h-auto gap-2 py-1.5 text-sm font-medium group-data-[disabled=true]/input-group:opacity-50 [&>kbd]:rounded-[calc(var(--radius)-5px)] [&>svg:not([class*='size-'])]:size-4 flex cursor-text items-center justify-center select-none",
-  {
-    defaultVariants: {
-      align: 'inline-start'
-    },
-    variants: {
-      align: {
-        'block-end': 'px-2.5 pb-2 group-has-[>input]/input-group:pb-2 [.border-t]:pt-2 order-last w-full justify-start',
-        'block-start':
-          'px-2.5 pt-2 group-has-[>input]/input-group:pt-2 [.border-b]:pb-2 order-first w-full justify-start',
-        'inline-end': 'pr-2 has-[>button]:mr-[-0.3rem] has-[>kbd]:mr-[-0.15rem] order-last',
-        'inline-start': 'pl-2 has-[>button]:ml-[-0.3rem] has-[>kbd]:ml-[-0.15rem] order-first'
-      }
-    }
-  }
-);
-
-//ensure that Group encapsulates the input and the addon
-const InputGroupAddon = ({
-  align = 'inline-start',
-  className,
-  ...props
-}: React.ComponentProps<'div'> & VariantProps<typeof inputGroupAddonVariants>) => {
-  return (
-    <div
-      className={cn(inputGroupAddonVariants({ align }), className)}
-      data-align={align}
-      data-slot="input-group-addon"
-      {...props}
-    />
-  );
-};
-
-const inputGroupButtonVariants = cva('gap-2 text-sm flex items-center shadow-none', {
-  defaultVariants: {
-    size: 'xs'
-  },
-  variants: {
-    size: {
-      'icon-sm': 'size-8 p-0 has-[>svg]:p-0',
-      'icon-xs': 'size-6 rounded-[calc(var(--radius)-3px)] p-0 has-[>svg]:p-0',
-      sm: '',
-      xs: "h-6 gap-1 rounded-[calc(var(--radius)-3px)] px-1.5 [&>svg:not([class*='size-'])]:size-3.5"
-    }
-  }
-});
-
-const InputGroupButton = ({
-  className,
-  size = 'xs',
-  type = 'button',
-  variant = 'ghost',
-  ...props
-}: Omit<React.ComponentProps<typeof Button>, 'size'> & VariantProps<typeof inputGroupButtonVariants>) => {
-  return (
-    <Button
-      className={cn(inputGroupButtonVariants({ size }), className)}
-      data-size={size}
-      type={type}
-      variant={variant}
-      {...props}
-    />
-  );
-};
-
-const InputGroupText = ({ className, ...props }: React.ComponentProps<'span'>) => {
-  return (
-    <span
-      className={cn(
-        "text-muted-foreground flex items-center gap-2 text-sm [&_svg]:pointer-events-none [&_svg:not([class*='size-'])]:size-4",
-        className
-      )}
-      {...props}
-    />
-  );
-};
-
-const InputGroupInput = ({ className, ...props }: React.ComponentProps<'input'>) => {
-  return (
-    <Input
-      className={cn(
-        'flex-1 rounded-none border-0 bg-transparent shadow-none ring-0 focus-visible:ring-0 disabled:bg-transparent aria-invalid:ring-0 dark:bg-transparent dark:disabled:bg-transparent',
-        className
-      )}
-      data-slot="input-group-control"
-      {...props}
-    />
-  );
-};
-
-const InputGroupTextArea = ({ className, ...props }: React.ComponentProps<'textarea'>) => {
-  return (
-    <TextArea
-      className={cn(
-        'flex-1 resize-none rounded-none border-0 bg-transparent py-2 shadow-none ring-0 focus-visible:ring-0 disabled:bg-transparent aria-invalid:ring-0 dark:bg-transparent dark:disabled:bg-transparent',
-        className
-      )}
-      data-slot="input-group-control"
-      {...props}
-    />
-  );
-};
-
-export { InputGroup, InputGroupAddon, InputGroupButton, InputGroupInput, InputGroupText, InputGroupTextArea };
+export { InputGroup };
